Make config_source aware of cli defaults
If a value comes from CLI defaults, it will now return :cli_default.

This supports verifying deprecations when we need to know
if a value was actually supplied from the CLI, or if it was
defaulted from CLI options.

Since we have all of the original sources still available,
this also makes it so that we don't keep a separate hash
for tracking config source - we're just checking in the original
config sources in order that matches merge priority.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>