Change 'prerelease' to 'channel'.
This adds support for deprecating a boolean flag
into a non-boolean value, and uses it to make `--prerelease`
deprecated in favor of `--channel current`.  By default,
`channel` is `stable`.

Separately, all deprecated options are now configured
to display at the end of the options list, instead of mixed into
the non-deprecated list.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>